Improve wave generator and LFO modulation for pitch, TVF and TVA
* Add support for lookup table in CPUROM for sine function (used for LFOs)
* Use new sine lookup table for sine generation in Wave Generator class
* Improve LFO TVA modulation with proper sine function and correct range
* Improve LFO TVF modulation by using the TVF lookup table from CPUROM
* Improve LFO pitch modulation by using the pitch lookup table from CPUROM

+++ libemusc/src/tva.h
@@ -60,8 +60,8 @@ private:
 
   ControlRom::LookupTables &_LUT;
 
-  uint8_t _accLFO1Depth;
-  uint8_t _accLFO2Depth;
+  int _accLFO1Depth;
+  int _accLFO2Depth;
 
   uint8_t _key;
   int _drumSet;
